requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
silence-tensorflow==1.2.2
six==1.16.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
